common-java:DuplicatedBlocks - Source files should not have any duplicated blocks
